feat: 더미데이터 분리
